refactor: reorganize share token generation
- Moved `generateShareToken` from the lib directory to a new utils file for better structure and maintainability.
- Updated import paths in `shareTokens.ts` to reflect this change, enhancing code organization.
