+++ src/utils/share-token.ts
